Refactor server-domain module
