+++ server-domain/src/main/java/io/joyoungc/domain/common/constant/ErrorCode.java
@@ -1,4 +1,4 @@
-package io.joyoungc.domain.constant;
+package io.joyoungc.domain.common.constant;
 
 /***
  * Created by Aiden Jeong on 2022.02.24
