+++ server-domain/src/main/java/io/joyoungc/domain/common/exception/ApplicationException.java
@@ -1,7 +1,7 @@
 package io.joyoungc.domain.common.exception;
 
 
-import io.joyoungc.domain.constant.ErrorCode;
+import io.joyoungc.domain.common.constant.ErrorCode;
 
 /***
  * Created by Aiden Jeong on 2022.02.24
